+++ src/apps/components/AppAlerts/useAppsFailedDeliveries.ts
@@ -1,35 +1,15 @@
 import { useUserPermissions } from "@dashboard/auth/hooks/useUserPermissions";
-import {
-  AppFailedPendingWebhooksQuery,
-  EventDeliveryStatusEnum,
-  PermissionEnum,
-  useAppFailedPendingWebhooksLazyQuery,
-} from "@dashboard/graphql";
+import { PermissionEnum, useAppFailedPendingWebhooksLazyQuery } from "@dashboard/graphql";
 import { useMemo } from "react";
 
-type Webhook = NonNullable<
-  NonNullable<AppFailedPendingWebhooksQuery["apps"]>["edges"][0]["node"]["webhooks"]
->[0];
+import { webhookFailedAttemptsCheck } from "./utils";
+
 interface AppsFailedDeliveries {
   hasFailed: boolean;
   fetchAppsWebhooks: () => void;
 }
 
 const requiredPermissions = [PermissionEnum.MANAGE_APPS];
-
-const hasFailedAttemptsCheck = (webhook: Webhook) =>
-  webhook.failedDelivers && webhook.failedDelivers?.edges?.length > 0;
-const hasFailedAttemptsInPendingCheck = (webhook: Webhook) => {
-  const preliminaryCheck = webhook.pendingDelivers && webhook.pendingDelivers?.edges?.length > 0;
-
-  if (!preliminaryCheck) return false;
-
-  return webhook.pendingDelivers?.edges.some(edge =>
-    edge.node?.attempts?.edges.some(
-      attempt => attempt.node.status === EventDeliveryStatusEnum.FAILED,
-    ),
-  );
-};
 
 export const useAppsFailedDeliveries = (): AppsFailedDeliveries => {
   const permissions = useUserPermissions();
@@ -44,7 +24,7 @@ export const useAppsFailedDeliveries = (): AppsFailedDeliveries => {
       data?.apps?.edges.some(
         app =>
           app.node.webhooks?.some(webhook => {
-            if (hasFailedAttemptsCheck(webhook) || hasFailedAttemptsInPendingCheck(webhook)) {
+            if (webhookFailedAttemptsCheck(webhook)) {
               return true;
             }
