+++ src/fragments/apps.ts
@@ -96,6 +96,7 @@ export const appListItemFragment = gql`
     permissions {
       ...AppPermission
     }
+    ...AppEventDeliveries
   }
 `;
 
@@ -121,5 +122,42 @@ export const webhookAttemptFragment = gql`
     status
     response
     responseStatusCode
+  }
+`;
+
+export const appEventDeliveriesFragment = gql`
+  fragment AppEventDeliveries on App {
+    webhooks {
+      failedDelivers: eventDeliveries(
+        first: 1
+        filter: { status: FAILED }
+        sortBy: { field: CREATED_AT, direction: DESC }
+      ) {
+        edges {
+          node {
+            id
+          }
+        }
+      }
+      pendingDelivers: eventDeliveries(
+        first: 1
+        filter: { status: PENDING }
+        sortBy: { field: CREATED_AT, direction: DESC }
+      ) {
+        edges {
+          node {
+            id
+            attempts(first: 6, sortBy: { field: CREATED_AT, direction: DESC }) {
+              edges {
+                node {
+                  id
+                  status
+                }
+              }
+            }
+          }
+        }
+      }
+    }
   }
 `;
